requirements.txt
groq==0.15.0
langchain-groq==0.2.0
PyPDF2==3.0.1
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-core==0.3.0
streamlit==1.28.0
python-dotenv==1.0.0
